Refactoring Dear ImGui backend
