+++ Echo.UserInterface/src/Backend/IApplication.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Echo.UserInterface.Backend;
+
+public interface IApplication : IDisposable
+{
+	TimeSpan UpdateDelay { get; }
+
+	string Label { get; }
+
+	void Initialize();
+
+	void Update();
+}
