+++ Echo.UserInterface/src/Backend/Native.cs
@@ -1,0 +1,138 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using ImGuiNET;
+
+namespace Echo.UserInterface.Backend;
+
+using static SDL2.SDL;
+
+static class Native
+{
+	[DllImport("SDL2", CallingConvention = CallingConvention.Cdecl)]
+	public static extern unsafe int SDL_RenderGeometryRaw(IntPtr renderer,
+														  IntPtr texture,
+														  float* xy,
+														  int xy_stride,
+														  int* color,
+														  int color_stride,
+														  float* uv,
+														  int uv_stride,
+														  int num_vertices,
+														  IntPtr indices,
+														  int num_indices,
+														  int size_indices);
+
+
+	[DllImport("SDL2", CallingConvention = CallingConvention.Cdecl)]
+	public static extern int SDL_RenderSetViewport(IntPtr renderer, IntPtr rect);
+
+	public static ImGuiKey SDL_KeycodeToImGuiKey(SDL_Keycode key) => key switch
+	{
+		SDL_Keycode.SDLK_TAB => ImGuiKey.Tab,
+		SDL_Keycode.SDLK_LEFT => ImGuiKey.LeftArrow,
+		SDL_Keycode.SDLK_RIGHT => ImGuiKey.RightArrow,
+		SDL_Keycode.SDLK_UP => ImGuiKey.UpArrow,
+		SDL_Keycode.SDLK_DOWN => ImGuiKey.DownArrow,
+		SDL_Keycode.SDLK_PAGEUP => ImGuiKey.PageUp,
+		SDL_Keycode.SDLK_PAGEDOWN => ImGuiKey.PageDown,
+		SDL_Keycode.SDLK_HOME => ImGuiKey.Home,
+		SDL_Keycode.SDLK_END => ImGuiKey.End,
+		SDL_Keycode.SDLK_INSERT => ImGuiKey.Insert,
+		SDL_Keycode.SDLK_DELETE => ImGuiKey.Delete,
+		SDL_Keycode.SDLK_BACKSPACE => ImGuiKey.Backspace,
+		SDL_Keycode.SDLK_SPACE => ImGuiKey.Space,
+		SDL_Keycode.SDLK_RETURN => ImGuiKey.Enter,
+		SDL_Keycode.SDLK_ESCAPE => ImGuiKey.Escape,
+		SDL_Keycode.SDLK_QUOTE => ImGuiKey.Apostrophe,
+		SDL_Keycode.SDLK_COMMA => ImGuiKey.Comma,
+		SDL_Keycode.SDLK_MINUS => ImGuiKey.Minus,
+		SDL_Keycode.SDLK_PERIOD => ImGuiKey.Period,
+		SDL_Keycode.SDLK_SLASH => ImGuiKey.Slash,
+		SDL_Keycode.SDLK_SEMICOLON => ImGuiKey.Semicolon,
+		SDL_Keycode.SDLK_EQUALS => ImGuiKey.Equal,
+		SDL_Keycode.SDLK_LEFTBRACKET => ImGuiKey.LeftBracket,
+		SDL_Keycode.SDLK_BACKSLASH => ImGuiKey.Backslash,
+		SDL_Keycode.SDLK_RIGHTBRACKET => ImGuiKey.RightBracket,
+		SDL_Keycode.SDLK_BACKQUOTE => ImGuiKey.GraveAccent,
+		SDL_Keycode.SDLK_CAPSLOCK => ImGuiKey.CapsLock,
+		SDL_Keycode.SDLK_SCROLLLOCK => ImGuiKey.ScrollLock,
+		SDL_Keycode.SDLK_NUMLOCKCLEAR => ImGuiKey.NumLock,
+		SDL_Keycode.SDLK_PRINTSCREEN => ImGuiKey.PrintScreen,
+		SDL_Keycode.SDLK_PAUSE => ImGuiKey.Pause,
+		SDL_Keycode.SDLK_KP_0 => ImGuiKey.Keypad0,
+		SDL_Keycode.SDLK_KP_1 => ImGuiKey.Keypad1,
+		SDL_Keycode.SDLK_KP_2 => ImGuiKey.Keypad2,
+		SDL_Keycode.SDLK_KP_3 => ImGuiKey.Keypad3,
+		SDL_Keycode.SDLK_KP_4 => ImGuiKey.Keypad4,
+		SDL_Keycode.SDLK_KP_5 => ImGuiKey.Keypad5,
+		SDL_Keycode.SDLK_KP_6 => ImGuiKey.Keypad6,
+		SDL_Keycode.SDLK_KP_7 => ImGuiKey.Keypad7,
+		SDL_Keycode.SDLK_KP_8 => ImGuiKey.Keypad8,
+		SDL_Keycode.SDLK_KP_9 => ImGuiKey.Keypad9,
+		SDL_Keycode.SDLK_KP_PERIOD => ImGuiKey.KeypadDecimal,
+		SDL_Keycode.SDLK_KP_DIVIDE => ImGuiKey.KeypadDivide,
+		SDL_Keycode.SDLK_KP_MULTIPLY => ImGuiKey.KeypadMultiply,
+		SDL_Keycode.SDLK_KP_MINUS => ImGuiKey.KeypadSubtract,
+		SDL_Keycode.SDLK_KP_PLUS => ImGuiKey.KeypadAdd,
+		SDL_Keycode.SDLK_KP_ENTER => ImGuiKey.KeypadEnter,
+		SDL_Keycode.SDLK_KP_EQUALS => ImGuiKey.KeypadEqual,
+		SDL_Keycode.SDLK_LCTRL => ImGuiKey.LeftCtrl,
+		SDL_Keycode.SDLK_LSHIFT => ImGuiKey.LeftShift,
+		SDL_Keycode.SDLK_LALT => ImGuiKey.LeftAlt,
+		SDL_Keycode.SDLK_LGUI => ImGuiKey.LeftSuper,
+		SDL_Keycode.SDLK_RCTRL => ImGuiKey.RightCtrl,
+		SDL_Keycode.SDLK_RSHIFT => ImGuiKey.RightShift,
+		SDL_Keycode.SDLK_RALT => ImGuiKey.RightAlt,
+		SDL_Keycode.SDLK_RGUI => ImGuiKey.RightSuper,
+		SDL_Keycode.SDLK_APPLICATION => ImGuiKey.Menu,
+		SDL_Keycode.SDLK_0 => ImGuiKey._0,
+		SDL_Keycode.SDLK_1 => ImGuiKey._1,
+		SDL_Keycode.SDLK_2 => ImGuiKey._2,
+		SDL_Keycode.SDLK_3 => ImGuiKey._3,
+		SDL_Keycode.SDLK_4 => ImGuiKey._4,
+		SDL_Keycode.SDLK_5 => ImGuiKey._5,
+		SDL_Keycode.SDLK_6 => ImGuiKey._6,
+		SDL_Keycode.SDLK_7 => ImGuiKey._7,
+		SDL_Keycode.SDLK_8 => ImGuiKey._8,
+		SDL_Keycode.SDLK_9 => ImGuiKey._9,
+		SDL_Keycode.SDLK_a => ImGuiKey.A,
+		SDL_Keycode.SDLK_b => ImGuiKey.B,
+		SDL_Keycode.SDLK_c => ImGuiKey.C,
+		SDL_Keycode.SDLK_d => ImGuiKey.D,
+		SDL_Keycode.SDLK_e => ImGuiKey.E,
+		SDL_Keycode.SDLK_f => ImGuiKey.F,
+		SDL_Keycode.SDLK_g => ImGuiKey.G,
+		SDL_Keycode.SDLK_h => ImGuiKey.H,
+		SDL_Keycode.SDLK_i => ImGuiKey.I,
+		SDL_Keycode.SDLK_j => ImGuiKey.J,
+		SDL_Keycode.SDLK_k => ImGuiKey.K,
+		SDL_Keycode.SDLK_l => ImGuiKey.L,
+		SDL_Keycode.SDLK_m => ImGuiKey.M,
+		SDL_Keycode.SDLK_n => ImGuiKey.N,
+		SDL_Keycode.SDLK_o => ImGuiKey.O,
+		SDL_Keycode.SDLK_p => ImGuiKey.P,
+		SDL_Keycode.SDLK_q => ImGuiKey.Q,
+		SDL_Keycode.SDLK_r => ImGuiKey.R,
+		SDL_Keycode.SDLK_s => ImGuiKey.S,
+		SDL_Keycode.SDLK_t => ImGuiKey.T,
+		SDL_Keycode.SDLK_u => ImGuiKey.U,
+		SDL_Keycode.SDLK_v => ImGuiKey.V,
+		SDL_Keycode.SDLK_w => ImGuiKey.W,
+		SDL_Keycode.SDLK_x => ImGuiKey.X,
+		SDL_Keycode.SDLK_y => ImGuiKey.Y,
+		SDL_Keycode.SDLK_z => ImGuiKey.Z,
+		SDL_Keycode.SDLK_F1 => ImGuiKey.F1,
+		SDL_Keycode.SDLK_F2 => ImGuiKey.F2,
+		SDL_Keycode.SDLK_F3 => ImGuiKey.F3,
+		SDL_Keycode.SDLK_F4 => ImGuiKey.F4,
+		SDL_Keycode.SDLK_F5 => ImGuiKey.F5,
+		SDL_Keycode.SDLK_F6 => ImGuiKey.F6,
+		SDL_Keycode.SDLK_F7 => ImGuiKey.F7,
+		SDL_Keycode.SDLK_F8 => ImGuiKey.F8,
+		SDL_Keycode.SDLK_F9 => ImGuiKey.F9,
+		SDL_Keycode.SDLK_F10 => ImGuiKey.F10,
+		SDL_Keycode.SDLK_F11 => ImGuiKey.F11,
+		SDL_Keycode.SDLK_F12 => ImGuiKey.F12,
+		_ => ImGuiKey.None
+	};
+}
