+++ Echo.UserInterface/src/EchoUserInterface.cs
@@ -1,10 +1,15 @@
 ﻿using System;
+using Echo.UserInterface.Backend;
 using ImGuiNET;
 
 namespace Echo.UserInterface;
 
-public sealed class Application : IDisposable
+public sealed class EchoUserInterface : IApplication
 {
+	public TimeSpan UpdateDelay { get; } = TimeSpan.FromSeconds(1f / 60f);
+
+	public string Label => "Echo User Interface";
+
 	public void Initialize() { }
 
 	public void Update()
